requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface==0.3.0
python-multipart==0.0.20
faiss-cpu==1.11.0
httpx==0.28.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.173.0
python-dotenv==1.1.1
pytz==2025.2
transformers==4.53.0
sentence-transformers==5.0.0
twilio
